Plex servers are now stored correctly

+++ src/Domain/Entities/Plex/PlexAccountServer.cs
@@ -2,7 +2,7 @@
 {
     public class PlexAccountServer
     {
-        public long PlexAccountId { get; set; }
+        public int PlexAccountId { get; set; }
         public virtual PlexAccount PlexAccount { get; set; }
         public int PlexServerId { get; set; }
         public virtual PlexServer PlexServer { get; set; }
